+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
 
+## patch
+* style nodes in `nodePickerField` that are navigated to via keyboard
+* enable deleting nodes from `nodePickerField` with backspace
+
+
 ## v0.2.4
 * update all node pickers to request 25 per page
 * fix `nodePickerField` double scrollbar and overflow issues